requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
openpyxl>=3.1.0
pandas>=2.0.0
google-generativeai>=0.7.0
aiofiles>=23.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
